% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beachmat_colSums}
\alias{beachmat_colSums}
\title{Compute column sums from matrix using beachmat}
\usage{
beachmat_colSums(dmat)
}
\arguments{
\item{dmat}{a matrix.}
}
\value{
A vector with column means.
}
\description{
Input is a numeric matrix of any of the types supported by beachmat.
}
