% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocksize.R
\name{blocksize}
\alias{blocksize}
\title{blocksize}
\usage{
blocksize(data, ram = get_ram())
}
\arguments{
\item{data}{matrix-like object.}

\item{ram}{the max amount of ram (in bytes) to use.}
}
\value{
Numeric value of the maximum number of rows.
}
\description{
Return the maximum number of rows to use based on the amount of ram memory.
}
\examples{
data <- matrix(NA, nrow = 100, ncol=1000)
blocksize(data, ram=1e6)
}
