% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\docType{methods}
\name{kmeans}
\alias{kmeans}
\alias{kmeans,SingleCellExperiment-method}
\alias{kmeans,LinearEmbeddingMatrix-method}
\alias{kmeans,DelayedMatrix-method}
\alias{kmeans,matrix-method}
\title{k-means for large single cell sequencing data}
\usage{
\S4method{kmeans}{SummarizedExperiment}(x, ...)

\S4method{kmeans}{SingleCellExperiment}(x, reduceMethod = "PCA",
  whichAssay = 1, ...)

\S4method{kmeans}{LinearEmbeddingMatrix}(x, ...)

\S4method{kmeans}{DelayedMatrix}(x, ...)

\S4method{kmeans}{matrix}(x, ...)
}
\arguments{
\item{x}{The object on which to run k-means.}

\item{...}{Arguments to pass to \code{\link[stats]{kmeans}}.}

\item{reduceMethod}{Name of dimensionality reduction results to use as input
to k-means.}

\item{whichAssay}{The assay to use as input to k-means. Used only if
\code{reduceMethod = "none"}.}
}
\value{
k-means output
}
\description{
This is a wrapper for stats::kmeans() for
large single cell sequencing data with the dimensionality
reduction results as input in in the reducedDim() slot.
}
\examples{
se <- SummarizedExperiment(matrix(rnorm(100), ncol=10))
kmeans(se, centers = 2, reduceMethod = "none")
}
