% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\docType{methods}
\name{mbkmeans}
\alias{mbkmeans}
\alias{mbkmeans,SingleCellExperiment-method}
\alias{mbkmeans,LinearEmbeddingMatrix-method}
\alias{mbkmeans,ANY-method}
\title{Mini_Batch k-means for large single cell sequencing data}
\usage{
\S4method{mbkmeans}{SummarizedExperiment}(x, ...)

\S4method{mbkmeans}{SingleCellExperiment}(x, reduceMethod = "PCA",
  whichAssay = 1, ...)

\S4method{mbkmeans}{LinearEmbeddingMatrix}(x, ...)

\S4method{mbkmeans}{ANY}(x, clusters, batch_size = blocksize(x),
  max_iters = 10, num_init = 1, init_fraction = 1,
  initializer = "kmeans++", early_stop_iter = 10, verbose = FALSE,
  CENTROIDS = NULL, tol = 1e-04, seed = 1)
}
\arguments{
\item{x}{The object on which to run Mini_Batch k-means.}

\item{...}{Arguments to pass to \code{\link[stats]{kmeans}}.}

\item{reduceMethod}{Name of dimensionality reduction results to use as input
to k-means.}

\item{whichAssay}{The assay to use as input to Mini_Batch k-means. Used only if
\code{reduceMethod = "none"}.}

\item{clusters}{the number of clusters}

\item{batch_size}{the size of the mini batches}

\item{max_iters}{the maximum number of clustering iterations}

\item{num_init}{number of times the algorithm will be run with different
centroid seeds}

\item{init_fraction}{percentage of data to use for the initialization
centroids (applies if initializer is \emph{kmeans++} ). Should be a float
number between 0.0 and 1.0.}

\item{initializer}{the method of initialization. One of \emph{kmeans++} and
\emph{random}. See details for more information}

\item{early_stop_iter}{continue that many iterations after calculation of the
best within-cluster-sum-of-squared-error}

\item{verbose}{either TRUE or FALSE, indicating whether progress is printed
during clustering}

\item{CENTROIDS}{a matrix of initial cluster centroids. The rows of the
CENTROIDS matrix should be equal to the number of clusters and the columns
should be equal to the columns of the data}

\item{tol}{a float number. If, in case of an iteration (iteration > 1 and
iteration < max_iters) 'tol' is greater than the squared norm of the
centroids, then kmeans has converged}

\item{seed}{integer value for random number generator (RNG)}
}
\value{
k-means output

a list with the following attributes: centroids, WCSS_per_cluster,
 best_initialization, iters_per_initialization
}
\description{
This is a wrapper for stats::kmeans() for
large single cell sequencing data with the dimensionality
reduction results as input in in the reducedDim() slot.
}
\details{
This function performs k-means clustering using mini batches.

\strong{kmeans++}: kmeans++ initialization. Reference :
http://theory.stanford.edu/~sergei/papers/kMeansPP-soda.pdf AND
http://stackoverflow.com/questions/5466323/how-exactly-does-k-means-work

\strong{random}: random selection of data rows as initial centroids
}
\examples{
se <- SummarizedExperiment(matrix(rnorm(100), ncol=10))
mbkmeans(se, clusters = 2, reduceMethod = "none")
sce <- SingleCellExperiment(matrix(rnorm(100), ncol=10))
mbkmeans(sce, clusters = 2, reduceMethod = "none")
x<-matrix(rnorm(100), ncol=10)
mbkmeans(x,clusters = 3)

}
\references{
https://github.com/mlampros/ClusterR
}
