% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/MbkmeansParam.R
\docType{class}
\name{MbkmeansParam-class}
\alias{MbkmeansParam-class}
\alias{show,MbkmeansParam-method}
\alias{MbkmeansParam}
\title{Mini-batch k-means clustering}
\usage{
MbkmeansParam(centers, ...)
}
\description{
Run the mini-batch k-means \code{\link{mbkmeans}} function with the specified
number of centers within \code{\link{clusterRows}} from the
\code{\link{bluster}} Bioconductor package.
}
\details{
This function is deprecated. Please use the \code{MbkmeansParam} function in
the \code{\link{bluster}} Bioconductor package.
}
