% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MbkmeansParam.R
\docType{class}
\name{MbkmeansParam}
\alias{MbkmeansParam}
\alias{show,MbkmeansParam-method}
\title{Mini-batch k-means clustering}
\usage{
MbkmeansParam(centers, ...)
}
\arguments{
\item{centers}{An integer scalar specifying the number of centers.
Alternatively, a function that takes the number of observations and returns the number of centers. 
Note, the \code{\link{mbkmeans}} function uses the argument \code{clusters} argument to represent this argument. 
However, we use \code{centers} to match}

\item{...}{Further arguments to pass to \code{\link{mbkmeans}}.}
}
\description{
Run the mini-batch k-means \code{\link{mbkmeans}} function with the specified
number of centers within \code{\link[bluster]{clusterRows}} from the
\code{bluster} Bioconductor package.
}
\details{

}
