% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MB_blocksize.R
\name{predict_MB_blocksize}
\alias{predict_MB_blocksize}
\title{predict_MB_blocksize}
\usage{
predict_MB_blocksize(data, block_size, clusters, batch_size, init_fraction,
  max_iters, initializer = "kmeans++")
}
\arguments{
\item{data}{numeric matrix or integer matrix or HDF5Matrix}

\item{block_size}{the number of rows of data}

\item{clusters}{the number of clusters}

\item{batch_size}{the size of the mini batches}

\item{init_fraction}{percentage of data to use for the initialization centroids (applies if initializer is \emph{kmeans++} ). Should be a float number between 0.0 and 1.0.}

\item{max_iters}{the maximum number of clustering iterations}

\item{initializer}{the method of initialization. One of \emph{kmeans++} and \emph{random}}
}
\value{
It returns a vector with the clusters
}
\description{
Predict Mini-batch-k-means by block for both matrix and HDF5Matrix
}
\section{blocksize}{
 \code{\link[mbkmeans]{blocksize}}
}

