% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_mini_batch}
\alias{predict_mini_batch}
\title{Predict_mini_batch}
\usage{
predict_mini_batch(data, CENTROIDS = NULL, fuzzy = FALSE,
  eps = 1e-06)
}
\arguments{
\item{data}{numeric matrix or integer matrix or HDF5Matrix}

\item{CENTROIDS}{a matrix of initial cluster centroids. The rows of the CENTROIDS matrix should be equal to the number of clusters and the columns should equal the columns of the data.}
}
\value{
it returns a vector with the clusters.
}
\description{
Prediction function for Mini-batch-k-means for both matrix and HDF5Matrix
}
\details{
This function takes the data and the output centroids and returns the clusters.
}
\references{
https://github.com/mlampros/ClusterR
}
