% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocksize.R
\name{blocksize}
\alias{blocksize}
\title{blocksize}
\usage{
blocksize(data)
}
\arguments{
\item{data}{numeric matrix or integer matrix or HDF5Matrix}
}
\value{
It returns a value of the maximum number of rows
}
\description{
Return the maximum number of rows to use based on the amount of ram memory
}
