% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{clusterRows}
\alias{clusterRows}
\title{Cluster rows of a matrix}
\description{
Cluster rows of a matrix-like object with a variety of algorithms.
}
\details{
This function is deprecated. Please use the \code{clusterRows} function in
the \code{\link{bluster}} Bioconductor package.
}
